requirements.txt
gradio>=4.47.0
fastapi>=0.110.0
uvicorn>=0.30.0

transformers>=4.44.2
torch>=2.2.0
sentencepiece>=0.2.0

google-api-python-client>=2.143.0
instaloader>=4.13
requests>=2.32.3
langdetect>=1.0.9

pandas>=2.2.2
openpyxl>=3.1.5
python-dotenv>=1.0.1
rapidfuzz>=3.9.6
